drop dependency on Puma 6 specific status.h header
Temporary solution. Code added to ccsp-cm-agent is currently
bypassing the CH HAL abstraction and using Intel specific types (and
functions exported by the HAL but not defined in the HAL API). More
cleanup is required, but this should be enough to at least build the
code without a direct dependency on the Puma 6 SDK.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>