optimise startup time ( OFW-870 ) ( OFW-3458 )
Multiple entries in config file taking longer to reflect on DUT.

Optimize number of dmcli calls by sorting the entries and parsing the
output of getvalues for an object.

Note that this relies on a corresponding change in dmcli to add
support for a command line option to disable Alias Manager.

cleanup CWMP object processing in config file

Fix for the crash, if queried object is not available.

fix zombie process is seen in CMAgent ( OFW-4720 )

[Issue]
Zombie process is observed in CMAgent during bootup.
This process is created because the popen function was not closed properly.

[Solution]
To resolve this issue, you need to perform a pclose after using popen.

[Testing Performed]
Checked all the process status after the bootup. No Zombie process was seen.

Signed-off-by: Govindaraja N B <gnb.ext@libertyglobal.com>