add comment that COSA_CM_DHCP_INFO must exactly match CMMGMT_CM_DHCP_INFO in cm_hal.h
It's bad that this struct defintion is duplicated between the CM HAL
and ccsp-cm-agent. As a temporary solution until a real fix is made,
add a warning to cosa_x_cisco_com_cablemodem_apis.h to document that
the struct definition must be kept in sync with the HAL version.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>